refactor: move all elasticsearch metrics to the elasticsearch layer (#1651)
