+++ src/rubrix/server/commons/models.py
@@ -1,0 +1,8 @@
+from enum import Enum
+
+
+class TaskStatus(str, Enum):
+    default = "Default"
+    edited = "Edited"  # TODO: DEPRECATE
+    discarded = "Discarded"
+    validated = "Validated"
