+++ src/rubrix/server/elasticseach/metrics/datasets.py
@@ -1,0 +1,8 @@
+# All metrics related to the datasets index
+from rubrix.server.elasticseach.metrics.base import TermsAggregation
+
+METRICS = {
+    "all_rubrix_workspaces": TermsAggregation(
+        id="all_rubrix_workspaces", field="owner.keyword"
+    )
+}
