+++ src/rubrix/server/services/metrics/__init__.py
@@ -1,0 +1,2 @@
+from .models import *
+from .service import *
