+++ src/rubrix/server/services/text_classification.py
@@ -22,7 +22,7 @@ from rubrix.server.apis.v0.models.commons.model import (
     EsRecordDataFieldNames,
     SortableField,
 )
-from rubrix.server.apis.v0.models.metrics.base import BaseMetric, BaseTaskMetrics
+from rubrix.server.apis.v0.models.metrics.base import BaseTaskMetrics
 from rubrix.server.apis.v0.models.text_classification import (
     CreationTextClassificationRecord,
     DatasetLabelingRulesMetricsSummary,
@@ -36,6 +36,7 @@ from rubrix.server.apis.v0.models.text_classification import (
     TextClassificationSearchResults,
 )
 from rubrix.server.errors.base_errors import MissingDatasetRecordsError
+from rubrix.server.services.metrics import BaseMetric
 from rubrix.server.services.search.model import SortConfig
 from rubrix.server.services.search.service import SearchRecordsService
 from rubrix.server.services.storage.service import RecordsStorageService
